Capitulo 7 - Adquirindo criaturas (Incompleto)

+++ app/src/main/java/com/example/devmon/model/domain/Creature.kt
@@ -1,7 +1,7 @@
 package com.example.devmon.model.domain
 
 data class Creature(
-    val number: Long,
+    val number: Int,
     val name: String,
     val imageUrl: String,
 )
